+++ packages/icon/src/types/icon.types.ts
@@ -1,11 +1,19 @@
-import { Props } from "@fortawesome/react-native-fontawesome";
 import { ComponentSize } from "@impulse-ui-native/theme";
+import { ComponentType } from "react";
 import { ViewStyle } from "react-native";
 
 export interface IconStyle extends ViewStyle {
   color?: string;
 }
 
-export interface IconProps extends Omit<Props, "size"> {
+export interface IconSvgProps {
+  width: number;
+  height: number;
+  color?: string;
+}
+
+export interface IconProps {
   size?: ComponentSize | number;
+  color?: string;
+  icon: ComponentType<IconSvgProps>;
 }