+++ packages/echo/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @impulse-ui-native/echo
 
+## 1.0.2
+
+### Patch Changes
+
+- Publishing layer manager types
+- Updated dependencies
+  - @impulse-ui-native/layer-manager-types@1.0.2
+  - @impulse-ui-native/core@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Removing unpublishede package from dependencies
+- Updated dependencies
+  - @impulse-ui-native/core@1.0.1
+  - @impulse-ui-native/layer-manager-types@1.0.1
+
 ## 1.0.0
 
 ### Major Changes