+++ packages/flyout/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @impulse-ui-native/flyout
 
+## 1.0.2
+
+### Patch Changes
+
+- Publishing layer manager types
+- Updated dependencies
+  - @impulse-ui-native/layer-manager-types@1.0.2
+  - @impulse-ui-native/safe-area-view@1.0.2
+  - @impulse-ui-native/typography@1.0.2
+  - @impulse-ui-native/theme@1.0.2
+  - @impulse-ui-native/core@1.0.2
+  - @impulse-ui-native/view@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Removing unpublishede package from dependencies
+- Updated dependencies
+  - @impulse-ui-native/core@1.0.1
+  - @impulse-ui-native/layer-manager-types@1.0.1
+  - @impulse-ui-native/safe-area-view@1.0.1
+  - @impulse-ui-native/theme@1.0.1
+  - @impulse-ui-native/typography@1.0.1
+  - @impulse-ui-native/view@1.0.1
+
 ## 1.0.0
 
 ### Major Changes