+++ packages/input/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @impulse-ui-native/input
 
+## 1.0.2
+
+### Patch Changes
+
+- Publishing layer manager types
+- Updated dependencies
+  - @impulse-ui-native/typography@1.0.2
+  - @impulse-ui-native/theme@1.0.2
+  - @impulse-ui-native/core@1.0.2
+  - @impulse-ui-native/icon@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Removing unpublishede package from dependencies
+- Updated dependencies
+  - @impulse-ui-native/core@1.0.1
+  - @impulse-ui-native/icon@1.0.1
+  - @impulse-ui-native/theme@1.0.1
+  - @impulse-ui-native/typography@1.0.1
+
 ## 1.0.0
 
 ### Major Changes